requirements.txt
requests
feedparser
beautifulsoup4
readability-lxml
lxml
newspaper3k
pillow
moviepy
snownlp
gTTS
python-dateutil
PyYAML